xpath BUGFIX too small flags member of restrictions structure (when, must)
To store LYS_XPATH_DEP flag, we need 16b, not just 8b

+++ src/tree_schema.h
@@ -1933,7 +1933,7 @@ struct lys_restr {
     const char *emsg;                /**< error-message (optional) */
     struct lys_ext_instance **ext;   /**< array of pointers to the extension instances */
     uint8_t ext_size;                /**< number of elements in #ext array */
-    uint8_t flags;                   /**< only one flag can be specified, #LYS_XPATH_DEP */
+    uint16_t flags;                  /**< only one flag can be specified, #LYS_XPATH_DEP */
 };
 
 /**
@@ -1945,7 +1945,7 @@ struct lys_when {
     const char *ref;                 /**< reference (optional) */
     struct lys_ext_instance **ext;   /**< array of pointers to the extension instances */
     uint8_t ext_size;                /**< number of elements in #ext array */
-    uint8_t flags;                   /**< only one flag can be specified, #LYS_XPATH_DEP */
+    uint16_t flags;                  /**< only one flag can be specified, #LYS_XPATH_DEP */
 };
 
 /**
